Extract `ChatComposite` browser test app to external folder (#2261)
