+++ change/@internal-react-composites-ffd71ca8-8000-4a97-8b03-f9af1376bb4b.json
@@ -1,0 +1,7 @@
+{
+  "type": "none",
+  "comment": "Reorganize internals of browser test folder",
+  "packageName": "@internal/react-composites",
+  "email": "82062616+prprabhu-ms@users.noreply.github.com",
+  "dependentChangeType": "none"
+}
